creates CSS files for each component, relocates components into dedicated directories, adds Figma designs for light and dark modes
