+++ src/index.css
@@ -1,0 +1,24 @@
+
+/* 
+
+LIGHT MODE:
+
+- #F5EEE6
+- #FFF8E3
+- #E6A4B4
+- #F3D7CA
+
+- #000000
+- #FFFFFF
+
+
+
+DARK MODE:
+
+- #1C1C1B
+- #FFC0CB
+- #CE4A7E
+
+- #000000
+- #FFFFFF
+*/
